+++ src/components/Forms/shared/FormInput/index.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import React, { useState } from 'react';
 import styled, { css } from 'styled-components';
 
-import FormError from '../FormError';
+import FormError from '../FormError/FormError';
 import FormLabel from '../FormLabel';
 import { color, font, fontSize, withThemes, mixins } from '../../../../styles';
